+++ lib/src/command/create_command.dart
@@ -20,6 +20,16 @@ class CreateCommand extends CommandBase {
         help:
             'The organization responsible for your new Flutter project, in reverse domain name notation. This string is used in Java package names and as prefix in the iOS bundle identifier. (defaults to \"com.example\")');
 
+    argParser.addOption('provider-system',
+        abbr: 'p',
+        allowed: ['flutter_modular', 'bloc_pattern'],
+        help: 'Create a flutter project using an specified provider system.');
+
+    argParser.addOption('state-management',
+        abbr: 'm',
+        allowed: ['mobx', 'flutter_bloc', 'rxdart'],
+        help: 'Create a flutter project using an specified state management.');
+
     argParser.addFlag('kotlin',
         abbr: 'k', negatable: false, help: 'use kotlin in Android project');
 
@@ -37,12 +47,15 @@ class CreateCommand extends CommandBase {
           'project name not passed for a create command', usage);
     } else {
       create(
-          argResults.rest.first,
-          argResults['description'],
-          argResults['org'],
-          argResults['kotlin'],
-          argResults['swift'],
-          argResults['androidx']);
+        argResults.rest.first,
+        argResults['description'],
+        argResults['org'],
+        argResults['kotlin'],
+        argResults['swift'],
+        argResults['androidx'],
+        argResults['state-management'],
+        argResults['provider-system'],
+      );
     }
   }
 }