+++ lib/src/command/generate_command.dart
@@ -3,6 +3,7 @@ import 'package:slidy/src/command/sub_command/generate_bloc.dart';
 import 'package:slidy/src/command/sub_command/generate_controller.dart';
 import 'package:slidy/src/command/sub_command/generate_model_sub_command.dart';
 import 'package:slidy/src/command/sub_command/generate_service_sub_command.dart';
+import 'package:slidy/src/command/sub_command/generate_store.dart';
 import 'package:slidy/src/command/sub_command/generate_test_sub_command.dart';
 
 class GenerateCommand extends CommandBase {
@@ -22,8 +23,10 @@ class GenerateCommand extends CommandBase {
     addSubcommand(GenerateWidgetAbbrSubCommand());
     addSubcommand(GenerateBlocSubCommand());
     addSubcommand(GenerateBlocAbbrSubCommand());
+    addSubcommand(GenerateStoreSubCommand());
     addSubcommand(GenerateControllerSubCommand());
     addSubcommand(GenerateControllerAbbrSubCommand());
+    addSubcommand(GenerateStoreAbbrSubCommand());
     addSubcommand(GenerateRepositorySubCommand());
     addSubcommand(GenerateRepositoryAbbrSubCommand());
     addSubcommand(GenerateServiceSubCommand());