+++ lib/src/templates/generator/widget_test_generator.dart
@@ -1,7 +1,8 @@
-String widgetTestGenerator(String name, String packageName, String import,
-    String module, String pathModule, bool isModular) {
+import 'package:slidy/src/utils/object_generate.dart';
+
+String widgetTestGenerator(ObjectGenerate obj) {
   String package;
-  if (isModular) {
+  if (obj.isModular) {
     package = r"import 'package:flutter_modular/flutter_modular_test.dart'";
   } else {
     package = r"import 'package:bloc_pattern/bloc_pattern_test.dart'";
@@ -12,22 +13,21 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 $package;
 
-import 'package:${packageName}/${import.replaceFirst("lib/", "").replaceAll("\\", "/")}';
+import 'package:${obj.packageName}/${obj.import.replaceFirst("lib/", "").replaceAll("\\", "/")}';
 
 main() {
-  testWidgets('${name}Widget has message', (WidgetTester tester) async {
-    await tester.pumpWidget(buildTestableWidget(${name}Widget()));
-    final textFinder = find.text('${name}');
+  testWidgets('${obj.name}Widget has message', (WidgetTester tester) async {
+    await tester.pumpWidget(buildTestableWidget(${obj.name}Widget()));
+    final textFinder = find.text('${obj.name}');
     expect(textFinder, findsOneWidget);
   });
 }
   ''';
 }
 
-String widgetTestGeneratorWithoutSuffix(String name, String packageName,
-    String import, String module, String modulePath, bool isModular) {
+String widgetTestGeneratorWithoutSuffix(ObjectGenerate obj) {
   String package;
-  if (isModular) {
+  if (obj.isModular) {
     package = r"import 'package:flutter_modular/flutter_modular_test.dart'";
   } else {
     package = r"import 'package:bloc_pattern/bloc_pattern_test.dart'";
@@ -37,12 +37,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 $package;
-import 'package:${packageName}/${import.replaceFirst("lib/", "").replaceAll("\\", "/")}';
+import 'package:${obj.packageName}/${obj.import.replaceFirst("lib/", "").replaceAll("\\", "/")}';
 
 main() {
-  testWidgets('${name} has message', (WidgetTester tester) async {
-    await tester.pumpWidget(buildTestableWidget(${name}()));
-    final textFinder = find.text('${name}');
+  testWidgets('${obj.name} has message', (WidgetTester tester) async {
+    await tester.pumpWidget(buildTestableWidget(${obj.name}()));
+    final textFinder = find.text('${obj.name}');
     expect(textFinder, findsOneWidget);
   });
 }