+++ lib/src/utils/object_generate.dart
@@ -1,0 +1,19 @@
+class ObjectGenerate {
+  final String name;
+  final String packageName;
+  final String import;
+  final String type;
+  final String module;
+  final String pathModule;
+  final bool isModular;
+
+  ObjectGenerate({
+    this.isModular,
+    this.type,
+    this.name,
+    this.packageName,
+    this.import,
+    this.module,
+    this.pathModule,
+  });
+}